admin/backend: Extract open primary fixture into shared test helper
Pulls the open-primary CVR setup + crossover/flip adjudications out of
app.tally_report_data.test.ts into seedOpenPrimaryWithAdjudications in
test/open_primary_fixture.ts so the same ballot landscape can be reused
across data, PDF, and CSV assertion layers without divergence. No
behavior change.
